+++ source/core/board/NstBoardBmcMarioParty7in1.cpp
@@ -80,8 +80,8 @@ namespace Nes
 
 					const byte data[2] =
 					{
-						exRegs[0],
-						exRegs[1]
+						static_cast<byte>(exRegs[0]),
+						static_cast<byte>(exRegs[1])
 					};
 
 					state.Begin( AsciiId<'B','M','P'>::V ).Begin( AsciiId<'R','E','G'>::V ).Write( data ).End().End();
