+++ source/core/board/NstBoardBtlMarioBaby.cpp
@@ -90,9 +90,9 @@ namespace Nes
 				{
 					const byte data[3] =
 					{
-						irq.Connected() ? 0x1 : 0x0,
-						irq.unit.count >> 0 & 0xFF,
-						irq.unit.count >> 8 & 0x7F
+						static_cast<byte>(irq.Connected() ? 0x1 : 0x0),
+						static_cast<byte>(irq.unit.count >> 0 & 0xFF),
+						static_cast<byte>(irq.unit.count >> 8 & 0x7F)
 					};
 
 					state.Begin( AsciiId<'B','M','B'>::V ).Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End().End();
