+++ source/core/board/NstBoardFfe.cpp
@@ -169,8 +169,8 @@ namespace Nes
 					const byte data[3] =
 					{
 						irq->unit.enabled != false,
-						irq->unit.count & 0xFF,
-						irq->unit.count >> 8
+						static_cast<byte>(irq->unit.count & 0xFF),
+						static_cast<byte>(irq->unit.count >> 8)
 					};
 
 					state.Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End();
