+++ source/core/board/NstBoardIremG101.cpp
@@ -89,8 +89,8 @@ namespace Nes
 				{
 					const byte data[2] =
 					{
-						regs[0],
-						regs[1]
+						static_cast<byte>(regs[0]),
+						static_cast<byte>(regs[1])
 					};
 
 					state.Begin( AsciiId<'I','G','1'>::V ).Begin( AsciiId<'R','E','G'>::V ).Write( data ).End().End();
